fix(esm): support oop reporter in esm mode
